pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "motor>=3.7.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "telegram>=0.0.1",
]
